change display capabilities to manage_categories
even though saving is restricted to manage_options. This is a bit of a quirk but should match the old implementation. In case there are people with manage_categories that want to see the form data but don't need to save. If this stays we should eventually build this into the UI.